+++ changelog.md
@@ -8,7 +8,7 @@ The versioning pattern is `major.minor.patch`.
 
 ---
 ## v0.1.3
-- added support for installation in Windows environments
+- added unit tests in `tests` folder.
 
 ## v0.1.2
 - added download_task_output parameter to `CloudClient.monitor_job()` to download stdout and stderr of each task when the task completes.